requirements.txt
# Core Framework
Flask==3.1.2
Werkzeug==3.1.2

# Database
SQLAlchemy==2.0.46

# Image Generation
qrcode[pil]==8.0
Pillow==10.1.0

# Security & Production
flask-limiter==3.8.1
python-dotenv==1.0.0

# Data Validation
marshmallow==3.21.3

# Logging & Monitoring
python-json-logger==2.0.8

# WSGI Server (for production)
gunicorn==21.2.0

# Windows-friendly production runner
waitress==2.1.2
 
# Postgres driver for migration/staging
psycopg2-binary==2.9.10
